+++ src/main/java/finalpairmatching/domain/Course.java
@@ -1,4 +1,4 @@
-package finalpairmatching;
+package finalpairmatching.domain;
 
 import java.util.stream.Stream;
