+++ src/main/java/finalpairmatching/domain/PairResult.java
@@ -1,4 +1,4 @@
-package finalpairmatching;
+package finalpairmatching.domain;
 
 import java.util.List;
