+++ src/main/java/finalpairmatching/domain/PairResults.java
@@ -1,4 +1,4 @@
-package finalpairmatching;
+package finalpairmatching.domain;
 
 import java.util.ArrayList;
 import java.util.List;
